chore: cargo fmt

+++ src-tauri/crates/harness/src/test_support.rs
@@ -1661,7 +1661,7 @@ impl WorkflowExecutionRepository for EmptyWorkflowExecutionRepository {
         _status: &str,
         _output_result: Option<&str>,
         _node_executions: Option<&str>,
-        _total_time_ms: Option<i64>,
+        _total_time_ms: Option<i32>,
     ) -> std::result::Result<bool, String> {
         Ok(true)
     }
